Scrum Burndown Tool
Build with Laravel 5.4, Vuejs 2.3 and has support for the Jira Agile
REST API

The default Jira chart does not give you alot of information, it only
shows the story points left in the sprint.
To squels this information gap, behold the new burndown, featured with
muiltiple burndown lines. In this way the Product Owner, Scrum master
and the Development team can see the real progress of the sprint.

Not all companies have the (expencive) Jira software tool, to make this
burndown usefull for as many scrum masters and mortals, the tool can
also be used manualy (although a bit less cool).